+++ include/stxxl/bits/containers/queue.h
@@ -45,7 +45,7 @@ STXXL_BEGIN_NAMESPACE
 //! \tparam AllocStr parallel disk allocation strategy, default is \c STXXL_DEFAULT_ALLOC_STRATEGY
 //! \tparam SizeType size data type, default is \c external_size_type
 template <class ValueType,
-          unsigned BlockSize = STXXL_DEFAULT_BLOCK_SIZE(ValueType),
+          size_t BlockSize = STXXL_DEFAULT_BLOCK_SIZE(ValueType),
           class AllocStr = STXXL_DEFAULT_ALLOC_STRATEGY,
           class SizeType = external_size_type>
 class queue : private noncopyable
